Factor out forms to extra components
